requirements.txt

streamlit
torch
torchvision
pillow
realesrgan
